Don't create new TaskRuns if PipelineRun timed out
The PipelineRun controller uses a combination of pipeline.timeout, pipeline.timeouts, and
pipeline.tasks[].timeout to determine the appropriate timeout for a new TaskRun or finally TaskRun.
The TaskRun is timed out by the TaskRun controller when it reaches its timeout.

Prior to this commit, the "getTaskRunTimeout" and "getFinallyTaskRunTimeout" functions were used to
determine the appropriate timeout for a new TaskRun or finally TaskRun. Pipeline level timeouts *should*
never be exceeded at the time a new TaskRun is created, but if this situation occurred, the PipelineRun
controller would create a new TaskRun with a 1 second timeout.

This commit renames "getTaskRunTimeout" to "getTaskRunTimeRemaining" and documents this function (and
likewise for "getFinallyTaskRunTimeout"). It also modifies these functions to return an error if there is
no time remaining; that is, if Pipeline Tasks or finally Tasks have already exceeded pipeline level
timeouts. The new behavior will allow bugs where the reconciler attempts to create a TaskRun for an
already-timed-out PipelineRun to be detected more easily.

Additional refactoring and handling of retries will happen in subsequent commits.