Don't schedule retries when PipelineRun stops
Prior to this commit, when a PipelineRun was in a "stopping" state (i.e. a TaskRun failed),
other TaskRuns would not complete their retries. However, TaskRuns with retries would also
not be considered "finished", so no finally Tasks would be scheduled, and the PipelineRun would hang.

This commit updates the PipelineRun reconciler to not allow TaskRuns to exhaust their retries when other
TaskRuns have failed, and documents this behavior.