Replace Container with Tekton-owned struct
This commit amends the definition of Task.Step to embed a newly-defined Container struct
instead of the Kubernetes Container struct. The only difference between the two structs is that resource
requirements are permitted to be any string, rather than just strings representing a valid resource quantity
such as "10Gi". This is necessary in order to implement parameteriziation of resource requirements.
The same changes are applied to StepTemplates, Sidecars and TaskRuns.StepOverrides.

Any Task yaml that worked before this change will continue to work after this change; however, the Go client
libraries will receive breaking changes.